minor restructure and convention change
