+++ src/main/java/com/datasonnet/spi/DataFormatService.java
@@ -48,7 +48,7 @@ public class DataFormatService {
         return Collections.unmodifiableList(plugins);
     }
 
-    public Optional<DataFormatPlugin> thatProduces(MediaType output, Class<?> target) {
+    public Optional<DataFormatPlugin> thatCanWrite(MediaType output, Class<?> target) {
         for (DataFormatPlugin plugin : plugins) {
             if (plugin.canWrite(output, target)) {
                 return Optional.of(plugin);
@@ -57,7 +57,7 @@ public class DataFormatService {
         return Optional.empty();
     }
 
-    public Optional<DataFormatPlugin> thatAccepts(Document<?> doc) {
+    public Optional<DataFormatPlugin> thatCanRead(Document<?> doc) {
         for (DataFormatPlugin plugin : plugins) {
             if (plugin.canRead(doc)) {
                 return Optional.of(plugin);
@@ -67,13 +67,13 @@ public class DataFormatService {
     }
 
     public <T> Document<T> mandatoryWrite(Value input, MediaType mediaType, Class<T> targetType) throws PluginException {
-        return thatProduces(mediaType, targetType)
+        return thatCanWrite(mediaType, targetType)
                 .orElseThrow(() -> new IllegalArgumentException("The output MediaType " + mediaType + " is not supported for class" + targetType))
                 .write(input, mediaType, targetType);
     }
 
     public ujson.Value mandatoryRead(Document<?> doc) throws PluginException {
-        return thatAccepts(doc)
+        return thatCanRead(doc)
                 .orElseThrow(() -> new IllegalArgumentException("The input MediaType " + doc.getMediaType() + " is not supported for class" + doc.getContent().getClass()))
                 .read(doc);
     }
